move url_to_str to converters
