+++ zyte_common_items/items/article.py
@@ -13,8 +13,11 @@ from zyte_common_items.components import (
     Image,
     Video,
 )
-from zyte_common_items.converters import to_metadata_optional
-from zyte_common_items.util import url_to_str
+from zyte_common_items.converters import (
+    to_metadata_optional,
+    url_to_str,
+    url_to_str_optional,
+)
 
 
 @attrs.define(kw_only=True)
@@ -115,7 +118,7 @@ class Article(Item):
     #:
     #: See also ``url``.
     canonicalUrl: Optional[str] = attrs.field(
-        default=None, converter=attrs.converters.optional(url_to_str), kw_only=True
+        default=None, converter=url_to_str_optional, kw_only=True
     )
 
     #: The main URL of the article page.
