+++ zyte_common_items/items/article_list.py
@@ -12,8 +12,11 @@ from zyte_common_items.components import (
     ListMetadata,
     ProbabilityMetadata,
 )
-from zyte_common_items.converters import to_metadata_optional
-from zyte_common_items.util import url_to_str
+from zyte_common_items.converters import (
+    to_metadata_optional,
+    url_to_str,
+    url_to_str_optional,
+)
 
 
 @attrs.define(kw_only=True)
@@ -44,7 +47,7 @@ class ArticleList(Item):
     #:
     #: See also ``url``.
     canonicalUrl: Optional[str] = attrs.field(
-        default=None, converter=attrs.converters.optional(url_to_str), kw_only=True
+        default=None, converter=url_to_str_optional, kw_only=True
     )
 
     #: List of article details found on the page.
@@ -125,5 +128,5 @@ class ArticleFromList(Item):
 
     #: Main URL.
     url: Optional[str] = attrs.field(
-        default=None, converter=attrs.converters.optional(url_to_str), kw_only=True
+        default=None, converter=url_to_str_optional, kw_only=True
     )
