+++ zyte_common_items/items/job_posting.py
@@ -11,8 +11,7 @@ from zyte_common_items.components import (
     JobLocation,
     Metadata,
 )
-from zyte_common_items.converters import to_metadata_optional
-from zyte_common_items.util import url_to_str
+from zyte_common_items.converters import to_metadata_optional, url_to_str
 
 
 @attrs.define(kw_only=True)
