+++ zyte_common_items/items/product.py
@@ -14,8 +14,11 @@ from zyte_common_items.components import (
     Gtin,
     Image,
 )
-from zyte_common_items.converters import to_metadata_optional
-from zyte_common_items.util import url_to_str
+from zyte_common_items.converters import (
+    to_metadata_optional,
+    url_to_str,
+    url_to_str_optional,
+)
 
 
 @attrs.define(kw_only=True)
@@ -62,7 +65,7 @@ class Product(Item):
     #:
     #: See also ``url``.
     canonicalUrl: Optional[str] = attrs.field(
-        default=None, converter=attrs.converters.optional(url_to_str), kw_only=True
+        default=None, converter=url_to_str_optional, kw_only=True
     )
 
     #: Color.
@@ -277,7 +280,7 @@ class ProductVariant(Item):
     #:
     #: See also ``url``.
     canonicalUrl: Optional[str] = attrs.field(
-        default=None, converter=attrs.converters.optional(url_to_str), kw_only=True
+        default=None, converter=url_to_str_optional, kw_only=True
     )
 
     #: Color.
@@ -390,5 +393,5 @@ class ProductVariant(Item):
     #:
     #: See also ``canonicalUrl``.
     url: Optional[str] = attrs.field(
-        default=None, converter=attrs.converters.optional(url_to_str), kw_only=True
+        default=None, converter=url_to_str_optional, kw_only=True
     )
