+++ zyte_common_items/items/real_estate.py
@@ -13,8 +13,7 @@ from zyte_common_items.components import (
     Image,
     RealEstateArea,
 )
-from zyte_common_items.converters import to_metadata_optional
-from zyte_common_items.util import url_to_str
+from zyte_common_items.converters import to_metadata_optional, url_to_str
 
 
 @attrs.define(kw_only=True)
